requirements.txt
Flask>=3.1,<4
waitress>=3.0,<4
pdfplumber>=0.11,<1
pypdf>=5,<6
python-docx>=1.1,<2
python-pptx>=1.0,<2
openpyxl>=3.1,<4
reportlab>=4.0,<5
pillow>=10,<13
pytesseract>=0.3,<1
pdf2image>=1.17,<2
extract-msg>=0.55,<1
